fix: compontentised stat card
